design: Modal Page safe-area 적용
