+++ src/components/add/AddTagFormRouteAsModal.tsx
@@ -1,5 +1,3 @@
-import { css, Theme } from '@emotion/react';
-
 import PortalWrapper from '~/components/common/PortalWrapper';
 import { MODAL_TYPE } from '~/constants/common';
 import useQueryParam from '~/hooks/common/useRouterQuery';
@@ -12,25 +10,7 @@ export default function AddTagFormRouteAsModal() {
 
   return (
     <PortalWrapper isShowing={query === MODAL_TYPE.addTag}>
-      <div css={wrapperCss}>
-        <TagPage />
-      </div>
+      <TagPage />
     </PortalWrapper>
   );
 }
-
-const wrapperCss = (theme: Theme) => css`
-  position: fixed;
-  top: 0;
-  /* 가로 가운데 정렬 */
-  left: 50%;
-  transform: translateX(-50%);
-
-  width: 100%;
-  max-width: ${theme.size.maxWidth};
-  height: 100%;
-  overflow-y: scroll;
-  background-color: ${theme.color.background};
-  padding: ${theme.size.layoutPadding};
-  z-index: 1000;
-`;
