+++ src/components/common/PortalWrapper.tsx
@@ -1,6 +1,9 @@
 import { PropsWithChildren } from 'react';
+import { css, Theme } from '@emotion/react';
 import { AnimatePresence } from 'framer-motion';
 import { createPortal } from 'react-dom';
+
+import { safeAreaBottomCss, safeAreaTopCss } from '~/pages/_layout';
 
 interface ModalWrapperProps {
   isShowing: boolean;
@@ -14,10 +17,34 @@ export default function PortalWrapper({
 
   return container ? (
     createPortal(
-      <AnimatePresence exitBeforeEnter>{isShowing && children}</AnimatePresence>,
+      <AnimatePresence exitBeforeEnter>
+        {isShowing && (
+          <div css={wrapperCss}>
+            <div css={safeAreaTopCss} />
+            {children}
+            <div css={safeAreaBottomCss} />
+          </div>
+        )}
+      </AnimatePresence>,
       container
     )
   ) : (
     <></>
   );
 }
+
+const wrapperCss = (theme: Theme) => css`
+  position: fixed;
+  top: 0;
+  /* 가로 가운데 정렬 */
+  left: 50%;
+  transform: translateX(-50%);
+
+  width: 100%;
+  max-width: ${theme.size.maxWidth};
+  height: 100%;
+  overflow-y: scroll;
+  background-color: ${theme.color.background};
+  padding: ${theme.size.layoutPadding};
+  z-index: 1000;
+`;
