+++ src/components/inspiration/InspirationViewAsModal.tsx
@@ -1,5 +1,3 @@
-import { css, Theme } from '@emotion/react';
-
 import PortalWrapper from '~/components/common/PortalWrapper';
 import { MODAL_TYPE } from '~/constants/common';
 import usePreventScroll from '~/hooks/common/usePreventScroll';
@@ -14,25 +12,7 @@ export default function InspirationViewAsModal() {
 
   return (
     <PortalWrapper isShowing={isInspirationViewModal}>
-      <div css={wrapperCss}>
-        <ContentPage />
-      </div>
+      <ContentPage />
     </PortalWrapper>
   );
 }
-
-const wrapperCss = (theme: Theme) => css`
-  position: fixed;
-  top: 0;
-  /* 가로 가운데 정렬 */
-  left: 50%;
-  transform: translateX(-50%);
-
-  width: 100%;
-  max-width: ${theme.size.maxWidth};
-  height: 100%;
-  overflow-y: scroll;
-  background-color: ${theme.color.background};
-  padding: ${theme.size.layoutPadding};
-  z-index: 1000;
-`;
